moved Pcontroller code into dashboard, move components around, bug fixs
