+++ client/src/App.js
@@ -2,9 +2,6 @@ import React, { Component } from 'react'
 import { BrowserRouter as Router, Route, Redirect, Link } from 'react-router-dom'
 import './App.css'
 import Home from './containers/Home'
-import StudentProfile from './components/StudentProfile'
-import StudentList from './components/StudentList'
-
 import ProfileController from './components/ProfileController'
 
 
@@ -67,16 +64,12 @@ class App extends Component {
 
       render() {
         return (
-            <BrowserRouter>
+            <Router>
               <div>
-              <TopNav />
-                <TopNav logout={this.logout} />
-                <MainNav />
                 <Route path="/" exact component={Home} />
-                <Route path="/profile" component={StudentProfile} />
-                <Route path="/dashboard" component={Dashboard} />
+                <ProfileController />
               </div>
-            </BrowserRouter>
+            </Router>
         );
       }
     }
